[BE] Change to graphql-yoga
